==> analysis/3.0-effetto-su-attivita-fisica.Rmd <==
---
title: "Effetto della quarantena sull'attivita' fisica"
author: "Anna Quaglieri"
date: "31/05/2020"
output: 
  html_document:
    toc: true
---

```{r setup, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
library(readr)
library(tidyverse)
library(googledrive)
library(here)
library(ggalluvial)
library(modelsummary)
library(gridExtra)
library(epitools)
library(readxl)
library(ggExtra)
```


```{r read-data}
risposte <- read_csv(file.path(here(),"data/risposte-con-regioni.csv"))
```

# 1. Qual e' stato l'effetto della quarantena sull'attitudine a praticare attivita' fisica?

```{r plot-si-no-pratica-prima-dopo}
tab <- data.frame(table(risposte$ex_pre_si_no, risposte$ex_durante_si_mp)) %>%
  dplyr::rename(`Prima` = Var1,
                `Durante` = Var2)

ggplot(as.data.frame(tab),
       aes(y = Freq, axis1 = `Prima`, axis2 = `Durante`)) +
  geom_alluvium(aes(fill=`Prima`)) +
  geom_stratum(width = 4/12) +
  geom_text(stat = "stratum", infer.label = TRUE, size=4) +
  scale_x_discrete(limits = c("Prima", "Durante"), position = "top") +
  scale_fill_brewer(type = "qual", palette = "Set1")+
  theme_minimal()+
  labs(y = "Numero persone", fill = "Praticavi attivita fisica'\nprima della quarantena?") +
 theme(axis.text.x = element_text(face="bold", color="black", 
                           size=16, angle=0),
       legend.title = element_text(size=15,face="bold"),
       legend.text = element_text(size=13))
```

```{r table-si-no-pratica-prima-dopo}
knitr::kable(tab)
```

```{r association-test-rr}
risposte <- risposte %>%
  mutate(ex_pre_si_no = factor(ex_pre_si_no, levels = c("Si", "No")), 
         ex_durante_si_mp = factor(ex_durante_si_mp, levels = c("Si", "No")), 
         ex_pre_si_no_dic = as.numeric(factor(ex_pre_si_no, 
                                              levels = c("Si", "No")))-1,
         ex_durante_si_no_dic = as.numeric(factor(ex_durante_si_mp, 
                                                  levels = c("Si", "No")))-1) %>%
  mutate(change = factor(abs(ex_durante_si_no_dic-ex_pre_si_no_dic), 
                         levels = c(0,1), labels = c("No change", "Change")))

# Success=1 should be in the right column which in our case is "No" excericse before quarantine
riskratio(table(risposte$ex_pre_si_no, risposte$change))
```

## Come interpretare i risultati sopra 

- **Le persone che hanno dichiarato di non fare attivita' fisica prima della quarantena hanno una probabilita' 4 volte piu' grande di cambiare attitudine rispetto a chi gia' faceva attivita' prima della quarantena**


- Nella stima di questo rischio non stiamo tenendo in conto di nessun altro fattore delle persone che hanno risposto al questionario (eta, abitazione, sesso etc..)

- Il rapporto (Risk Ratio) si ottiene facendo il rapporto tra la probabilita' di cambiare tra chi ha detto **No** prima e chi ha detto **Si** prima: 

$$Risk\;Ratio_{No\;vs\;Si\;prima}=\frac{\frac{73}{124}}{\frac{109}{795}}=4.29$$
$$Risk\;Ratio_{Si\;vs\;No\;prima}=\frac{1}{Risk\;Ratio_{No\;vs\;Si\;prima}}=0.23$$


```{r prop-test-prima-dopo-pratica}
numero_persone <- c(929, 929)
yes <- c(sum(risposte$ex_pre_si_no == "Si"), sum(risposte$ex_durante_si_mp == "Si"))
prop.test(yes, numero_persone)
```

Nel complesso, la proporzione delle persone che pratica attivita' fisica diminuisce marginalmente, passando dall'`r round(prop.test(yes, numero_persone)$estimate[1],2)`% al `r round(prop.test(yes, numero_persone)$estimate[2],2)`%.


# 1.1 Tipi di attivita' fisica praticati prima e durante la quarantena


```{r tipi-attivita-prima-durante}
attivita_prima <- read_excel(file.path(here(), "data/dati_da_controllare_cambiati/tipi_esercizi_pre.xlsx")) %>%
  dplyr::select(Nomi_iniziali, `NOMI NUOVI`) %>% dplyr::rename(nomi_nuovi=`NOMI NUOVI`)
attivita_durante <- read_excel(file.path(here(), "data/dati_da_controllare_cambiati/tipi_esercizi_durante.xlsx")) %>%
  dplyr::select(Nomi_iniziali, `NOMI NUOVI`) %>% dplyr::rename(nomi_nuovi=`NOMI NUOVI`)
```

```{r merge-con-risposte, eval=FALSE}
risposte$id1 <- 1
risposte1 <- risposte %>%
  mutate(ex_durante = tolower(ex_durante),
         ex_pre = tolower(ex_pre)) %>%
  left_join(attivita_prima, by = c("ex_pre"="Nomi_iniziali")) %>%
  dplyr::select(-ex_pre) %>%
  dplyr::rename(ex_pre = nomi_nuovi) %>%
  left_join(attivita_durante, by = c("ex_durante"="Nomi_iniziali")) %>%
  dplyr::select(-ex_durante) %>%
  dplyr::rename(ex_durante = nomi_nuovi)
```


# 2. Motivazioni che hanno portato a un cambiamento dell'attitudine a praticare attivita' fisica.

## Persone che hanno smesso di praticare: motivazioni

```{r ricodifica-motivazioni}
motivi_smesso_pratica <- read_excel(file.path(here(), "data/dati_da_controllare_cambiati/motivo_ex_durante_no.xlsx")) %>%
  dplyr::select(Nomi_iniziali, `NOMI NUOVI`) %>% dplyr::rename(nomi_nuovi=`NOMI NUOVI`)
motivi_non_praticanti_prima <- read_excel(file.path(here(), "data/dati_da_controllare_cambiati/motivo_ex_pre_no.xlsx")) %>%
  dplyr::select(Nomi_iniziali, `nomi nuovi`) %>% dplyr::rename(nomi_nuovi=`nomi nuovi`)
motivi_esercizi_durante <- read_excel(file.path(here(), "data/dati_da_controllare_cambiati/motivo_ex_durante.xlsx")) %>%
  dplyr::select(Nomi_iniziali, `NOMI NUOVI`) %>% dplyr::rename(nomi_nuovi=`NOMI NUOVI`)
```

* Metti a posto abitazione e stato professionale

```{r abitazione}
casa <- read_excel(file.path(here(), "data/dati_da_controllare_cambiati/tipo_abitazione.xlsx")) %>%
  dplyr::select(Nomi_iniziali, `NOMI NUOVI`) %>% dplyr::rename(nomi_nuovi=`NOMI NUOVI`)
professione <- read_excel(file.path(here(), "data/dati_da_controllare_cambiati/stato_prof.xlsx")) %>%
  dplyr::select(Nomi_iniziali, `NOMI NUOVI`) %>% dplyr::rename(nomi_nuovi=`NOMI NUOVI`)
```

```{r motivazioni-si-no}
cambio_si_no <- risposte %>%
  filter(ex_pre_si_no == "Si" & ex_durante_si_mp == "No") %>%
  mutate(ex_durante_no_motivo = tolower(ex_durante_no_motivo)) %>%
  # Merge motivi
  left_join(motivi_smesso_pratica, by = c("ex_durante_no_motivo"="Nomi_iniziali")) %>%
  dplyr::select(-ex_durante_no_motivo) %>%
  dplyr::rename(ex_durante_no_motivo = nomi_nuovi) %>%
  group_by(ex_durante_no_motivo) %>%
  summarise(N.Risposte = n()) %>%
  arrange(N.Risposte) %>%
  mutate(ex_durante_no_motivo = fct_reorder(ex_durante_no_motivo, N.Risposte))

knitr::kable(cambio_si_no)
```

```{r}
ggplot(cambio_si_no, aes(x=ex_durante_no_motivo,y=N.Risposte)) + geom_bar(stat = "identity") +
  theme_bw() + coord_flip() + labs(x = "Motivazione che ha portato a smettere di praticare")
```


```{r demographics-si-no}
cambio_si_no <- risposte %>%
  filter(ex_pre_si_no == "Si" & ex_durante_si_mp == "No") %>%
  mutate(tipo_abitazione = tolower(tipo_abitazione)) %>%
  left_join(casa, by = c("tipo_abitazione"="Nomi_iniziali")) %>%
  dplyr::select(-tipo_abitazione) %>%
  dplyr::rename(tipo_abitazione = nomi_nuovi) %>%
  mutate(praticanti = "No")

prop.table(table(cambio_si_no$ex_pre_dove))
prop.table(table(cambio_si_no$ex_pre_frequenza))
```

* Il tipo di abitazione o stato professionale e' diverso in quelli che hanno smesso di praticare rispetto agli altri?

```{r}
praticanti <- risposte %>%
  filter(ex_durante_si_mp == "Si") %>%
  mutate(tipo_abitazione = tolower(tipo_abitazione)) %>%
  left_join(casa, by = c("tipo_abitazione"="Nomi_iniziali")) %>%
  dplyr::select(-tipo_abitazione) %>%
  dplyr::rename(tipo_abitazione = nomi_nuovi) %>%
  mutate(praticanti = "Si")
```

* Lo stato professionale non e' stato raccolto per quelli che hanno smesso di praticare attivita'! Per sbaglio..

## Cos'e' cambiato per quelli che hanno smesso rispetto a quelli che hanno continuato

### Abitazione

```{r eta-praticanti-e-non}

prat <- nrow(praticanti)
non_prat <- nrow(cambio_si_no)
combine_si_no <- bind_rows(cambio_si_no, praticanti) %>%
  group_by(tipo_abitazione, praticanti) %>%
  summarise(n = n()) %>%
  mutate(percent_praticanti = case_when(praticanti == "No" ~ n/non_prat * 100,
                                        praticanti == "Si" ~ n/prat * 100))

ggplot(combine_si_no, aes(x = tipo_abitazione, y = percent_praticanti, fill = praticanti)) + 
  geom_bar(position="dodge", stat = "identity") + theme_bw() +
  ggtitle("Confronto tra i tipo di abitazione di chi ha smesso di praticare\nattivita' fisica (praticanti=No) e chi ha continuato/iniziato durante la quarantena (praticanti=Si)") + labs(x = "Tipo abitazione", y = "Percentuale partecipanti per gruppo") + coord_flip()
```

Ci sono `r prat` partecipanti che hanno continuato/iniziato a praticare e `r non_prat` che hanno smesso. 

### Stato professionale non raccolto

### Luogo della pratica

```{r aperto-chiuso-praticanti-e-non}
praticanti <- risposte %>%
  filter(ex_pre_si_no == "Si" & ex_durante_si_mp == "Si") %>%
  mutate(praticanti = "Si")

prat <- nrow(praticanti)
non_prat <- nrow(cambio_si_no)
combine_si_no <- bind_rows(cambio_si_no, praticanti) %>%
  group_by(ex_pre_dove, praticanti) %>%
  summarise(n = n()) %>%
  mutate(percent_praticanti = case_when(praticanti == "No" ~ n/non_prat * 100,
                                        praticanti == "Si" ~ n/prat * 100))

combine_si_no

ggplot(combine_si_no, aes(x = ex_pre_dove, y = percent_praticanti, fill = praticanti)) + 
  geom_bar(position="dodge", stat = "identity") + theme_bw() +
  ggtitle("Confronto tra luogo dell'attivita' tra chi ha smesso di praticare\nattivita' fisica (praticanti=No) e chi ha continuato durante la quarantena (praticanti=Si)")+ labs(x = "Luogo dell'attivita' fisica prima della quarantena", y = "Percentuale partecipanti per gruppo")
```

Ci sono `r prat` partecipanti che hanno continuato a praticare e `r non_prat` che hanno smesso. 

### Metodo della pratica

```{r metodo-praticanti-e-non}
praticanti <- risposte %>%
  filter(ex_pre_si_no == "Si" & ex_durante_si_mp == "Si") %>%
  mutate(praticanti = "Si")

prat <- nrow(praticanti)
non_prat <- nrow(cambio_si_no)
combine_si_no <- bind_rows(cambio_si_no, praticanti) %>%
  group_by(ex_pre_come, praticanti) %>%
  summarise(n = n()) %>%
  mutate(percent_praticanti = case_when(praticanti == "No" ~ n/non_prat * 100,
                                        praticanti == "Si" ~ n/prat * 100))

combine_si_no

ggplot(combine_si_no, aes(x = ex_pre_come, y = percent_praticanti, fill = praticanti)) + 
  geom_bar(position="dodge", stat = "identity") + theme_bw() +
  ggtitle("Confronto tra metodo dell'attivita' tra chi ha smesso di praticare\nattivita' fisica (praticanti=No) e chi ha continuato durante la quarantena (praticanti=Si)")+ labs(x = "Metodo di attivita' fisica prima della quarantena", y = "Percentuale partecipanti per gruppo")
```

Ci sono `r prat` partecipanti che hanno continuato a praticare e `r non_prat` che hanno smesso. 


### Personal trainer?

```{r personal-trainer-praticanti-e-non}
praticanti <- risposte %>%
  filter(ex_pre_si_no == "Si" & ex_durante_si_mp == "Si") %>%
  mutate(praticanti = "Si")

prat <- nrow(praticanti)
non_prat <- nrow(cambio_si_no)
combine_si_no <- bind_rows(cambio_si_no, praticanti) %>%
  group_by(ex_pre_pt, praticanti) %>%
  summarise(n = n()) %>%
  mutate(percent_praticanti = case_when(praticanti == "No" ~ n/non_prat * 100,
                                        praticanti == "Si" ~ n/prat * 100))

combine_si_no

ggplot(combine_si_no, aes(x = ex_pre_pt, y = percent_praticanti, fill = praticanti)) + 
  geom_bar(position="dodge", stat = "identity") + theme_bw() +
  ggtitle("Confronto tra metodo dell'attivita' tra chi ha smesso di praticare\nattivita' fisica (praticanti=No) e chi ha continuato durante la quarantena (praticanti=Si)") + labs(x = "Attivita' fisica con personal trainer prima della quarantena?", y = "Percentuale partecipanti per gruppo")
```

Ci sono `r prat` partecipanti che hanno continuato a praticare e `r non_prat` che hanno smesso. 


## Persone che hanno iniziato a praticare: motivazioni

**Motivazioni per non praticare prima della quarantena**

```{r pre-motivazioni-no-si}
cambio_no_si_pre <- risposte %>%
  filter(ex_pre_si_no == "No" & ex_durante_si_mp == "Si") %>%
  mutate(ex_pre_no_motivo = tolower(ex_pre_no_motivo)) %>%
  left_join(motivi_non_praticanti_prima, by = c("ex_pre_no_motivo"="Nomi_iniziali")) %>%
  dplyr::select(-ex_pre_no_motivo) %>%
  dplyr::rename(ex_pre_no_motivo = nomi_nuovi) %>%
  #mutate(ex_pre_no_motivo = ifelse(!(ex_pre_no_motivo %in% c("Mancanza di voglia", "Mancanza di tempo")), "Altro", ex_pre_no_motivo)) %>%
  group_by(ex_pre_no_motivo) %>%
  summarise(N.Risposte = n()) %>%
  arrange(N.Risposte) %>%
  mutate(ex_pre_no_motivo = fct_reorder(ex_pre_no_motivo, N.Risposte)) 

knitr::kable(cambio_no_si_pre)
```


```{r}
ggplot(cambio_no_si_pre, aes(x=ex_pre_no_motivo,y=N.Risposte)) + geom_bar(stat = "identity") +
  theme_bw() + coord_flip() + labs(x = "Motivazione per non praticare prima della quarantena")
```

**Motivazioni per praticare durante la quarantena**

```{r include=FALSE}
cambio_no_si <- risposte %>%
  filter(ex_pre_si_no == "No" & ex_durante_si_mp == "Si")
```

* Alcune persone hanno messo piu' di una motivazione per iniziare a praticare. Il grafico sotto mostra la frequenza delle varie motivazioni. In tutto ci sono `r nrow(cambio_no_si)` persone che hanno iniziato a praticare. 

```{r durante-motivazioni-no-si}
cambio_no_si <- risposte %>%
  filter(ex_pre_si_no == "No" & ex_durante_si_mp == "Si") %>%
  mutate(ex_durante_motivo = tolower(ex_durante_motivo)) 


lista_motivazioni <- data.frame(ex_durante_motivo = tolower(do.call(c,strsplit(cambio_no_si$ex_durante_motivo,split=";")))) %>%
  left_join(motivi_esercizi_durante, by = c("ex_durante_motivo"="Nomi_iniziali")) %>%
  dplyr::select(-ex_durante_motivo) %>%
  dplyr::rename(ex_durante_motivo = nomi_nuovi) %>%
  group_by(ex_durante_motivo) %>%
  summarise(N.Risposte = n()) %>%
  arrange(N.Risposte) %>%
  mutate(ex_durante_motivo = fct_reorder(ex_durante_motivo, N.Risposte))
knitr::kable(lista_motivazioni)
```

```{r}
ggplot(lista_motivazioni, aes(x=ex_durante_motivo,y=N.Risposte)) + geom_bar(stat = "identity") +
  theme_bw() + coord_flip() + labs(x = "Motivazione per iniziare a praticare durante la quarantena")
```


# 3. Qual e' stato l'effetto della quarantena sulla frequenza della pratica di attivita' fisica?

```{r all-plot-frequenza}
risposte <- risposte %>%
  mutate(ex_pre_frequenza = ifelse(is.na(ex_pre_frequenza), "Non praticante", ex_pre_frequenza),
         ex_durante_frequenza = ifelse(is.na(ex_durante_frequenza), "Non praticante", ex_durante_frequenza))
tab <- data.frame(table(risposte$ex_pre_frequenza, risposte$ex_durante_frequenza)) %>%
  dplyr::rename(Prima = Var1,
                Durante = Var2) %>%
  mutate(Prima = factor(Prima, levels = c("Tutti i giorni", "Almeno 3 volte a settimana",
                                          "1 o 2 volte a settimana","Meno di una volta a settimana","Non praticante"), 
                        labels = c("Ogni giorno","Almeno 3", "1 o 2", "Meno di una", "Non\npraticante")),
         Durante = factor(Durante, levels = c("Tutti i giorni", "Almeno 3 volte a settimana",
                                          "1 o 2 volte a settimana","Meno di una volta a settimana","Non praticante"), 
                        labels = c("Ogni giorno","Almeno 3", "1 o 2", "Meno di una", "Non\npraticante")))

ggplot(as.data.frame(tab),
       aes(y = Freq, axis1 = `Prima`, axis2 = `Durante`)) +
  geom_alluvium(aes(fill=`Prima`)) +
  geom_stratum(width = 2/12,) +
  geom_text(stat = "stratum", infer.label = TRUE) +
  scale_x_discrete(limits = c("Prima", "Durante"), expand = c(0.03, .03), position = "top") +
  scale_fill_brewer(type = "qual", palette = "Set1")+
  theme_minimal()+
  labs(y = "Numero persone", fill = "Risposta prima della quarantena\n(Relativa alla frequenza per settimana)") +
  theme(legend.position = "bottom",
        axis.text.x = element_text(face="bold", color="black", 
                           size=16, angle=0),
       legend.title = element_text(size=13,face="bold"),
       legend.text = element_text(size=13))+
  guides(fill=guide_legend(nrow=2,byrow=TRUE))
```


```{r}
tab
```

```{r barplot-frequenza}
tab_long <- risposte %>%
  gather(key = Periodo, value = Frequenza, ex_pre_frequenza,ex_durante_frequenza)  %>%
  mutate(Periodo = ifelse(Periodo == "ex_durante_frequenza", "Durante", "Prima"))

ggplot(tab_long, aes(x = Frequenza, fill = Periodo)) +
  geom_bar(position = "dodge", colour="white") +
  labs(x = "Frequenza per settimana", y = "Numero di persone")+
  theme_bw() + coord_flip()
```


* Converti in numeri

  - Non praticante = 0
  - Meno di una volta a settimana = 1,
  - 1 o 2 volte a settimana = 2,
  - Almeno 3 volte a settimana = 3, 
  - Ogni giorno = 4

```{r convert-freq-numeri}
tab_numeri <- tab %>%
  mutate(Prima = case_when(Prima %in% "Non\npraticante" ~ 0, 
                           Prima %in% "Meno di una" ~ 1,
                           Prima %in% "1 o 2" ~ 2,
                           Prima %in% "Almeno 3" ~ 3, 
                           Prima %in% "Ogni giorno" ~ 4),
         Durante = case_when(Durante %in% "Non\npraticante" ~ 0, 
                           Durante %in% "Meno di una" ~ 1,
                           Durante %in% "1 o 2" ~ 2,
                           Durante %in% "Almeno 3" ~ 3, 
                           Durante %in% "Ogni giorno" ~ 4)) %>%
  mutate(Cambiamento = Durante - Prima)

ggplot(tab, aes(x = Prima, y = Durante, size = Freq)) + geom_point() + theme_bw() 
```

Il grafico sotto mostra la distribuzione della differenza tra la frequenza nella pratica di attivita' fisica prima e dopo. In generale c'e' un numero simile di persone che aumentano o diminuiscono la frequenza anche se leggeremente piu' persone aumentano la frequenza (la distribuzione e' piuttosto simmetrica attorno allo zero, media = 0.041 e t-test per l'uguaglianza della media a zero non e' significativo al 5%).

```{r cambiamento-prima-dopo}
tab_numeri <- risposte %>%
  mutate(Prima = case_when(ex_pre_frequenza %in% "Non praticante" ~ 0, 
                           ex_pre_frequenza %in% "Meno di una volta a settimana" ~ 1,
                           ex_pre_frequenza %in% "1 o 2 volte a settimana" ~ 2,
                           ex_pre_frequenza %in% "Almeno 3 volte a settimana" ~ 3, 
                           ex_pre_frequenza %in% "Tutti i giorni" ~ 4),
         Durante = case_when(ex_durante_frequenza %in% "Non praticante" ~ 0, 
                           ex_durante_frequenza %in% "Meno di una volta a settimana" ~ 1,
                           ex_durante_frequenza %in% "1 o 2 volte a settimana" ~ 2,
                           ex_durante_frequenza %in% "Almeno 3 volte a settimana" ~ 3, 
                           ex_durante_frequenza %in% "Tutti i giorni" ~ 4)) %>%
  mutate(Cambiamento = Durante - Prima)

ggplot(tab_numeri, aes(x = Cambiamento)) + geom_bar() + theme_bw() + labs(x = "Cambiamento\n(Durante - Prima)")
```

```{r t-test-prima-dopo-frequenza}
t.test(tab_numeri$Cambiamento)
```

# 5. Quelli che hanno praticato attivita' durante la quarantena, percepiscono un miglioramento nella loro forma fisica?

```{r miglioramento}
risposte <- read_csv(file.path(here(),"data/risposte-rename.csv"))
risposte <- risposte %>%
  mutate(forma_fisica_migliorata = ifelse(is.na(forma_fisica_migliorata), "Non praticante", as.character(forma_fisica_migliorata))) %>%
  mutate(forma_fisica_migliorata = factor(forma_fisica_migliorata, 
                                          levels = c("Non praticante", "Per niente", "Ho notato un lieve miglioramento", "Molto")))


ggplot(risposte, aes(x = ex_pre_si_no, fill = forma_fisica_migliorata)) +
  geom_bar(position = "dodge", colour="white")+
  theme_bw()+
  labs(x = "Praticavi attivita' fisica\nprima della quarantena?", y = "Numero di persone")
```



```{r}
risposte <- risposte %>%
  filter(forma_fisica_migliorata != "Non praticante")
tot_si <- sum(risposte$ex_pre_si_no == "Si")
tot_no <- sum(risposte$ex_pre_si_no == "No")
miglioramento_per_gruppo <- risposte %>% 
  group_by(forma_fisica_migliorata, ex_pre_si_no) %>%
  summarise(N.partecipanti = n()) %>%
  mutate(N.partecipanti = case_when(ex_pre_si_no == "Si" ~ N.partecipanti/tot_si * 100,
                                    ex_pre_si_no == "No" ~ N.partecipanti/tot_no *100))

ggplot(miglioramento_per_gruppo, aes(x = forma_fisica_migliorata, y = N.partecipanti, colour = ex_pre_si_no)) +
  geom_point(size=4) + theme_bw() + 
  labs(x = "Miglioramento nella forma fisica", y = "Percentuale\nCalcolata per gruppi", colour = "Praticavi prima\ndella quarantena?")

knitr::kable(miglioramento_per_gruppo)
```


# 6. Per quelli che praticavano anche prima della quarantena, come si sono adattati alla situazione?

```{r}
praticanti <- risposte %>%
  filter(ex_pre_si_no == "Si" & ex_durante_si_mp == "Si") %>%
  mutate(ex_durante_dove = ifelse(ex_durante_dove == "All'aperto (ad esempio giardino)", "All'aperto", ex_durante_dove))
```


## Luogo dell'attivita' 

```{r}
ggplot(praticanti, aes(x= ex_pre_dove, fill = ex_durante_dove)) + geom_bar(position="fill") + theme_bw()+
  labs(x = "Luogo prima della quarantena", fill = "Luogo durante la quarantena") 
```

## Metodo dell'attivita' 

```{r}
ggplot(praticanti, aes(x= ex_pre_come, fill = ex_durante_come)) + geom_bar(position="fill") + theme_bw()+
  labs(x = "Metodo prima della quarantena", fill = "Metodo durante la quarantena") 
```


## Personal trainer

```{r}
ggplot(praticanti, aes(x= ex_pre_pt, fill = ex_durante_pt)) + geom_bar(position="fill") + theme_bw()+
  labs(x = "Personal trainer prima della quarantena?", fill = "Personal trainer durante la quarantena?") 
```


# 6. Modello congiutno con tutti i fattori
