==> analysis/4.0-benessere-umorale.Rmd <==
---
title: "Effetto della quarantena e attivita' fisica sul benessere umorale"
author: "Anna Quaglieri"
date: "31/05/2020"
output: 
  html_document:
    toc: true
---

```{r setup, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(echo = FALSE, message = FALSE, warning = FALSE)
library(readr)
library(tidyverse)
library(googledrive)
library(here)
library(ggalluvial)
library(modelsummary)
library(gridExtra)
library(epitools)
library(readxl)
library(ggExtra)

library(ggmap)
library(maps)
library(mapproj)
library(viridis)
library(ggrepel)
```


```{r read-data}
risposte <- read_csv(file.path(here(),"data/risposte-con-regioni.csv"))
```

* Crea nuova variabile per cambio in attivita'

```{r}
risposte <- risposte %>%
  mutate(pratica = case_when(ex_pre_si_no == "Si" & ex_durante_si_mp == "Si" ~ "continuato",
                             ex_pre_si_no == "Si" & ex_durante_si_mp == "No" ~ "smesso",
                             ex_pre_si_no == "No" & ex_durante_si_mp == "No" ~ "mai praticato",
                             ex_pre_si_no == "No" & ex_durante_si_mp == "Si" ~ "iniziato"), 
         pratica_durante = case_when(ex_pre_si_no == "Si" & ex_durante_si_mp == "Si" ~ "Si",
                             ex_pre_si_no == "Si" & ex_durante_si_mp == "No" ~ "No",
                             ex_pre_si_no == "No" & ex_durante_si_mp == "No" ~ "No",
                             ex_pre_si_no == "No" & ex_durante_si_mp == "Si" ~ "Si"), ) %>%
  mutate(pratica = factor(pratica, levels = c("continuato", "smesso", "mai praticato", "iniziato")))
```

# Analisi univariata

Inizialmente ho ricodificato le risposte sullo stato umorale in numeri per vedere se c'e' stata una diminuzione o aumento generale. La ricodifica e' stata fatta in questo modo:

- Mai = 0
- Raramente = 1
- Ogni tanto = 2
- Spesso = 3
- Ogni giorno = 4

```{r numeric-variables}
umori <- c("ansia", "tristezza", "stress", "rabbia", "felicita", "entusiasmo", "calma", "soddisfazione")
var_pre <- paste0(umori, "_pre")
var_durante <- paste0(umori, "_durante")
positive <- c("felicita", "entusiasmo", "calma", "soddisfazione")
negative <- c("ansia", "tristezza", "stress", "rabbia")

risposte_var_numeric <- risposte %>%
  mutate_at(.vars = c(var_pre, var_durante),
      funs(case_when(
    . == "Mai" ~ 0,
    . == "Raramente" ~ 1,
     . == "Ogni tanto" ~ 2,
     . == "Spesso" ~ 3,
     . == "Ogni giorno" ~ 4)))
```

## C'e' stata una diminuzione o aumento in generale dei diversi tipi di umore?

```{r differenze}
risposte_numeric_long <- risposte_var_numeric %>%
  mutate(ansia = ansia_durante - ansia_pre,
         tristezza = tristezza_durante - tristezza_pre,
         stress = stress_durante - stress_pre,
         rabbia = rabbia_durante - rabbia_pre,
         felicita = felicita_durante - felicita_pre,
         entusiasmo = entusiasmo_durante - entusiasmo_pre,
         calma = calma_durante - calma_pre,
         soddisfazione = soddisfazione_durante - soddisfazione_pre) %>%
  pivot_longer(ansia:soddisfazione, names_to = "Umore", values_to = "differenze_durante_prima") 

risposte_numeric_long <- risposte_numeric_long %>%
  mutate(classe_umore = ifelse(Umore %in% c("felicita", "entusiasmo",
                                               "calma", "soddisfazione"), "Umori positivi","Umori negativi"))
```


Dato che stiamo guardando alla differenza tra prima e dopo, si avranno valori negativi se **prima** della quarantena i partecipanti hanno messo valori piu' alti rispetto al **durante** per certi aspetti umorali e viceversa, valori positivi se **durante** e' maggiore di **prima**.

```{r grafico_differenza}
ggplot(risposte_numeric_long, aes(x = differenze_durante_prima)) + geom_bar() +
  facet_wrap(~Umore) + theme_classic()+ scale_x_continuous(breaks = seq(-4,4,1)) + labs(x = "Differenza in umore: Durante - Prima")+
  ggtitle("Valori positivi = la sensazione e' aumetata durante la quarantena\nValori negativi = la sensazione e' diminuita durante la quarantena")+
  geom_vline(xintercept = 0, linetype="dotted", colour="dark red")
```

```{r grafico_differenza_class}
ggplot(risposte_numeric_long, aes(x = differenze_durante_prima, fill = Umore)) + geom_bar() +
  facet_wrap(~classe_umore) + theme_classic() + 
  scale_x_continuous(breaks = seq(-4,4,1)) + labs(x = "Differenza in umore: Durante - Prima")+
  ggtitle("Valori positivi = la sensazione e' aumetata durante la quarantena\nValori negativi = la sensazione e' diminuita durante la quarantena")
```


## La differenza e' stata influenzata dalla pratica di attivita'?

In particolare dall'attivita' fisica. Nei grafici sotto, invece che mostrare i numeri totali delle persone mostro la proporzione dei livelli dei gruppi (diversi colori) nelle varie barre. 

```{r umore_pratica}
ggplot(risposte_numeric_long, aes(x = differenze_durante_prima, fill = pratica)) + geom_bar(position = "fill") +
  facet_wrap(~Umore) + theme_classic()+ scale_x_continuous(breaks = seq(-4,4,1)) + labs(x = "Differenza in umore: Durante - Prima")+
  ggtitle("Valori positivi = la sensazione e' aumetata durante la quarantena\nValori negativi = la sensazione e' diminuita durante la quarantena")
```

```{r umore_pratica_fill}
ggplot(risposte_numeric_long, aes(x = differenze_durante_prima, fill = pratica_durante)) + geom_bar(position = "fill") +
  facet_wrap(~Umore) + theme_classic()+ scale_x_continuous(breaks = seq(-4,4,1)) + labs(x = "Differenza in umore: Durante - Prima")+
  ggtitle("Valori positivi = la sensazione e' aumetata durante la quarantena\nValori negativi = la sensazione e' diminuita durante la quarantena")
```

```{r umore_pratica_01}
ggplot(risposte_numeric_long, aes(x = differenze_durante_prima, fill = pratica_durante)) + geom_bar() +
  facet_wrap(~Umore) + theme_classic()+ scale_x_continuous(breaks = seq(-4,4,1)) + labs(x = "Differenza in umore: Durante - Prima")+
  ggtitle("Valori positivi = la sensazione e' aumetata durante la quarantena\nValori negativi = la sensazione e' diminuita durante la quarantena")
```

## Modelli 

```{r mulitin-models}
library(nnet)

risposte_numeric <- risposte_var_numeric %>%
  mutate(ansia = ansia_durante - ansia_pre,
         tristezza = tristezza_durante - tristezza_pre,
         stress = stress_durante - stress_pre,
         rabbia = rabbia_durante - rabbia_pre,
         felicita = felicita_durante - felicita_pre,
         entusiasmo = entusiasmo_durante - entusiasmo_pre,
         calma = calma_durante - calma_pre,
         soddisfazione = soddisfazione_durante - soddisfazione_pre)
```


- Per ogni sensazione fai modello lineare con umore durante e pratica di attivita' e tipo abitazione, sesso, eta
- Per ogni sensazione fai modello multinomial con prima e dopo e pratica di attivita' (e tipo abitazione), sesso, eta


1. Relevel di tutti gli umori mettendo la classe `Mai` come baseline

```{r relevel-umori}
risposte <- risposte %>%
  mutate_at(.vars = c(var_pre, var_durante), 
            .funs = funs(factor(. , levels = c("Mai", "Raramente", "Ogni tanto", "Spesso", "Ogni giorno"))))

risposte_long <- risposte %>%
  pivot_longer(ansia_pre:soddisfazione_durante, names_to = "umore", values_to = "livello") %>%
  mutate(classe_umore = ifelse(umore %in% c("felicita_durante", "entusiasmo_durante",
                                               "calma_durante", "soddisfazione_durante"), "Umori positivi (Durante)","Umori negativi (Durante)"))
```

2. Fai un modello per ogni umore e attivita' praticata durante aggiustando per sesso ed eta
  - Estrai: coefficienti e 95CI per il RR per i coefficienti della pratica durante
  
  
```{r function-extract-confint}
confint_and_est <- function(fit){
  bind_cols(RR = exp(coef(fit))[,"ex_durante_si_mpSi"], 
            t(confint(fit)[2,,]),
            livello =  rownames(t(confint(fit)[2,,]))) %>%
    mutate(low.conf = exp(`2.5 %`), high.conf = exp(`97.5 %`)) %>%
    select(-`2.5 %`, -`97.5 %`)
}

confint_and_est_eta <- function(fit){
  bind_cols(RR = exp(coef(fit))[,"ex_durante_si_mpSi"], 
            t(confint(fit)[2,,]),
            livello =  rownames(t(confint(fit)[2,,]))) %>%
    mutate(low.conf = exp(`2.5 %`), high.conf = exp(`97.5 %`)) %>%
    select(-`2.5 %`, -`97.5 %`)
}

```

### Modello con tutti i dati
  
```{r modello-nested}
modello <- risposte_long %>%
  filter(umore %in% var_durante) %>%
  dplyr::select(livello, ex_durante_si_mp, Sesso, eta, umore) %>%
  nest(-umore) %>%
  mutate(fit = purrr::map(data, ~ multinom(livello ~ ex_durante_si_mp + Sesso + eta, data = .))) %>%
  mutate(results_fit = purrr::map(fit, function(f) confint_and_est(f))) %>%
  mutate(z_score = purrr::map(fit, function(x) summary(x)$coefficients/summary(x)$standard.errors)) %>%
  mutate(pval = purrr::map(z_score, function(z) (1 - pnorm(abs(z), 0, 1)) * 2)) 

espandi_modelli <- modello %>%
  unnest(results_fit, .drop=TRUE) %>%
  mutate(livello = factor(livello, levels = c("Raramente", "Ogni tanto", "Spesso", "Ogni giorno")))
```
  
3. Plotta tutti i RR per tutti gli umori

```{r ansia-durante-attivita}
ggplot(espandi_modelli) +
  geom_point(aes(x=livello, y = RR)) +
  geom_line(aes(x=livello, y = RR, group = umore), colour = "dark red", linetype="dotted") +
  geom_errorbar(aes(x=livello, y = RR, ymin=low.conf, ymax=high.conf), width=0.2, size=0.6) +
  theme_bw() +
  facet_wrap(~umore) +
  geom_hline(yintercept = 1, linetype = "dotted")
```


### Combine positive/negative feelings

```{r combine-pos-neg}
modello <- risposte_long %>%
  filter(umore %in% var_durante) %>%
  dplyr::select(livello, ex_durante_si_mp, Sesso, eta, classe_umore) %>%
  nest(-classe_umore) %>%
  mutate(fit = purrr::map(data, ~ multinom(livello ~ ex_durante_si_mp + Sesso + eta, data = .))) %>%
  mutate(results_fit = purrr::map(fit, function(f) confint_and_est(f))) %>%
  mutate(z_score = purrr::map(fit, function(x) summary(x)$coefficients/summary(x)$standard.errors)) %>%
  mutate(pval = purrr::map(z_score, function(z) (1 - pnorm(abs(z), 0, 1)) * 2)) 

espandi_modelli <- modello %>%
  unnest(results_fit, .drop=TRUE) %>%
  mutate(livello = factor(livello, levels = c("Raramente", "Ogni tanto", "Spesso", "Ogni giorno")))
```

```{r combine-pos-neg-plot}
ggplot(espandi_modelli) +
  geom_point(aes(x=livello, y = RR)) +
  geom_line(aes(x=livello, y = RR, group = classe_umore), colour = "dark red", linetype="dotted") +
  geom_errorbar(aes(x=livello, y = RR, ymin=low.conf, ymax=high.conf), width=0.2, size=0.6) +
  theme_bw() +
  facet_wrap(~classe_umore) +
  geom_hline(yintercept = 1, linetype = "dotted")+
  labs(y = "logOR", x = "Frequenza percepita")
```

I LogOddsRatio per ogni livello sono calcolati con rispetto a chi pratica attivita' sportiva durante la quarantena o chi non pratica. 

### Modello con solo quelli che prima non praticavano

```{r modello-prina-no-attivita}
modello <- risposte_long %>%
  filter(umore %in% var_durante & ex_pre_si_no == "No") %>%
  dplyr::select(livello, ex_durante_si_mp, Sesso, eta, classe_umore) %>%
  nest(-classe_umore) %>%
  mutate(fit = purrr::map(data, ~ multinom(livello ~ ex_durante_si_mp + Sesso + eta, data = .))) %>%
  mutate(results_fit = purrr::map(fit, function(f) confint_and_est(f))) %>%
  mutate(z_score = purrr::map(fit, function(x) summary(x)$coefficients/summary(x)$standard.errors)) %>%
  mutate(pval = purrr::map(z_score, function(z) (1 - pnorm(abs(z), 0, 1)) * 2)) 

espandi_modelli <- modello %>%
  unnest(results_fit, .drop=TRUE) %>%
  mutate(livello = factor(livello, levels = c("Raramente", "Ogni tanto", "Spesso", "Ogni giorno")))

ggplot(espandi_modelli) +
  geom_point(aes(x=livello, y = RR)) +
  geom_line(aes(x=livello, y = RR, group = classe_umore), colour = "dark red", linetype="dotted") +
  geom_errorbar(aes(x=livello, y = RR, ymin=low.conf, ymax=high.conf), width=0.2, size=0.6) +
  theme_bw() +
  facet_wrap(~classe_umore) +
  geom_hline(yintercept = 1, linetype = "dotted") +
  labs(y = "logOR", x = "Frequenza percepita")
```

### Modello con solo quelli che prima praticavano

```{r modello-prina-si-attivita}
modello <- risposte_long %>%
  filter(umore %in% var_durante & ex_pre_si_no == "Si") %>%
  dplyr::select(livello, ex_durante_si_mp, Sesso, eta, classe_umore) %>%
  nest(-classe_umore) %>%
  mutate(fit = purrr::map(data, ~ multinom(livello ~ ex_durante_si_mp + Sesso + eta, data = .))) %>%
  mutate(results_fit = purrr::map(fit, function(f) confint_and_est(f))) %>%
  mutate(z_score = purrr::map(fit, function(x) summary(x)$coefficients/summary(x)$standard.errors)) %>%
  mutate(pval = purrr::map(z_score, function(z) (1 - pnorm(abs(z), 0, 1)) * 2)) 

espandi_modelli <- modello %>%
  unnest(results_fit, .drop=TRUE) %>%
  mutate(livello = factor(livello, levels = c("Raramente", "Ogni tanto", "Spesso", "Ogni giorno")))

ggplot(espandi_modelli) +
  geom_point(aes(x=livello, y = RR)) +
  geom_line(aes(x=livello, y = RR, group = classe_umore), colour = "dark red", linetype="dotted") +
  geom_errorbar(aes(x=livello, y = RR, ymin=low.conf, ymax=high.conf), width=0.2, size=0.6) +
  theme_bw() +
  facet_wrap(~classe_umore) +
  geom_hline(yintercept = 1, linetype = "dotted")+
  labs(y = "logOR", x = "Frequenza percepita")
```


### Modello influenza umore da sesso e eta

```{r modello-eta}
confint_and_est_eta <- function(fit){
  confint_Eta36_55 = data.frame(t(confint(fit)[3,,]))
  colnames(confint_Eta36_55) <- c("Eta36_55.low_conf", "Eta36_55.high_conf")
  confint_Eta36_55 <- confint_Eta36_55 %>%
    mutate(Eta36_55.low_conf = exp(Eta36_55.low_conf), Eta36_55.high_conf = exp(Eta36_55.high_conf))
    
  
  confint_Oltre_55 = data.frame(t(confint(fit)[4,,]))
  colnames(confint_Oltre_55) <- c("EtaOltre_55.low_conf", "EtaOltre_55.high_conf")
  confint_Oltre_55 <- confint_Oltre_55 %>%
    mutate(EtaOltre_55.low_conf = exp(EtaOltre_55.low_conf), 
           EtaOltre_55.high_conf = exp(EtaOltre_55.high_conf))
    
  
  bind_cols(RR_eta36_55 = exp(coef(fit))[,"etaDai 36 ai 55 anni"],
            RR_Oltre_55 = exp(coef(fit))[,"etaOltre i 55 anni"],
            livello = rownames(t(confint(fit)[4,,])),
            confint_Eta36_55,confint_Oltre_55) 
}

modello <- risposte_long %>%
  filter(umore %in% var_durante & ex_pre_si_no == "Si") %>%
  dplyr::select(livello, ex_durante_si_mp, Sesso, eta, classe_umore) %>%
  nest(-classe_umore) %>%
  mutate(fit = purrr::map(data, ~ multinom(livello ~ Sesso + eta, data = .))) %>%
  mutate(results_fit = purrr::map(fit, function(f) confint_and_est_eta(f))) %>%
  mutate(z_score = purrr::map(fit, function(x) summary(x)$coefficients/summary(x)$standard.errors)) %>%
  mutate(pval = purrr::map(z_score, function(z) (1 - pnorm(abs(z), 0, 1)) * 2)) 

espandi_low_conf <- modello %>%
  unnest(results_fit, .drop=TRUE) %>%
  mutate(livello = factor(livello, levels = c("Raramente", "Ogni tanto", "Spesso", "Ogni giorno"))) %>%
  gather(key = classe_eta, value = low.conf, Eta36_55.low_conf, EtaOltre_55.low_conf) %>%
  dplyr::select(classe_umore, livello, classe_eta, low.conf) %>%
  mutate(classe_eta = str_remove(classe_eta, ".low_conf"))

espandi_high_conf <- modello %>%
  unnest(results_fit, .drop=TRUE) %>%
  mutate(livello = factor(livello, levels = c("Raramente", "Ogni tanto", "Spesso", "Ogni giorno"))) %>%
  gather(key = classe_eta, value = high.conf, Eta36_55.high_conf, EtaOltre_55.high_conf) %>%
  dplyr::select(classe_umore, livello, classe_eta, high.conf) %>%
  mutate(classe_eta = str_remove(classe_eta, ".high_conf"))

espandi_RR <- modello %>%
  unnest(results_fit, .drop=TRUE) %>%
  mutate(livello = factor(livello, levels = c("Raramente", "Ogni tanto", "Spesso", "Ogni giorno"))) %>%
  gather(key = classe_eta, value = RR, RR_eta36_55, RR_Oltre_55) %>%
  dplyr::select(classe_umore, livello, classe_eta, RR) %>%
  mutate(classe_eta = case_when(classe_eta %in% "RR_eta36_55" ~ "Eta36_55",
                                classe_eta %in% "RR_Oltre_55" ~ "EtaOltre_55"))
  
espandi_modelli <- espandi_low_conf %>%
  left_join(espandi_high_conf) %>%
  left_join(espandi_RR)  

ggplot(espandi_modelli, aes(x = livello, y = RR)) +
  geom_pointrange(aes(ymin = low.conf, ymax = high.conf, colour = classe_eta), 
                    position=position_jitter(width=0.5)) +
  facet_wrap(~ classe_umore)+
  theme_bw() +
  geom_hline(yintercept = 1, linetype = "dotted")
```


```{r modello-sesso}
confint_and_est_sesso <- function(fit){
  confint_sesso = data.frame(t(confint(fit)[2,,]))
  colnames(confint_sesso) <- c("SessoMaschio_low.conf", "SessoMaschio_high.conf")
  confint_sesso <- confint_sesso %>%
    mutate(SessoMaschio_low.conf = exp(SessoMaschio_low.conf),
           SessoMaschio_high.conf = exp(SessoMaschio_high.conf))
  
  bind_cols(RR = exp(coef(fit))[,"SessoMaschio"],
            livello = rownames(t(confint(fit)[2,,])),
            confint_sesso) 
}

modello <- risposte_long %>%
  filter(umore %in% var_durante & ex_pre_si_no == "Si") %>%
  dplyr::select(livello, ex_durante_si_mp, Sesso, eta, classe_umore) %>%
  nest(-classe_umore) %>%
  mutate(fit = purrr::map(data, ~ multinom(livello ~ Sesso + eta, data = .))) %>%
  mutate(results_fit = purrr::map(fit, function(f) confint_and_est_sesso(f))) %>%
  mutate(z_score = purrr::map(fit, function(x) summary(x)$coefficients/summary(x)$standard.errors)) %>%
  mutate(pval = purrr::map(z_score, function(z) (1 - pnorm(abs(z), 0, 1)) * 2)) 

espandi_modelli <- modello %>%
  unnest(results_fit, .drop=TRUE) %>%
  mutate(livello = factor(livello, levels = c("Raramente", "Ogni tanto", "Spesso", "Ogni giorno"))) 

ggplot(espandi_modelli, aes(x = livello, y = RR)) +
  geom_pointrange(aes(ymin = SessoMaschio_low.conf, ymax = SessoMaschio_high.conf)) +
  facet_wrap(~ classe_umore)+
  theme_bw() +
  geom_hline(yintercept = 1, linetype = "dotted")
```


# Distribuzione geografica delle sensazioni 

Mostra la mappa dell'Italia con le varie sensazioni durante e la differenza tra prima e dopo


```{r load-geo}
# Get the world polygon and extract UK
italy <- map_data("world") %>% 
  filter(region=="Italy")

cities <- world.cities %>% 
  filter(country.etc=="Italy") %>%
  dplyr::rename(paese_residenza=name) %>%
  mutate(paese_residenza=tolower(paese_residenza))

# Regioni
province <- map_data("italy")
```

## In generale umori positivi vs umori negativi

```{r plot-geo, fig.height=12, fig.width=12}
risposte_by_city <- risposte %>%
  left_join(cities, by = c("comune" = "paese_residenza")) %>%
  group_by(comune, lat, long) %>%
  summarise(numero_risposte = n())

mybreaks <- round(seq(
  min(risposte_by_city$numero_risposte),
  max(risposte_by_city$numero_risposte), length.out = 5))

risposte_numeric_long <- risposte_numeric_long %>%
  left_join(cities, by = c("comune" = "paese_residenza")) %>%
  left_join(risposte_by_city)

ggplot() +
  geom_polygon(data = italy, aes(x= long, y = lat, group = group), fill="grey", alpha=0.3) +
  geom_point(data=risposte_numeric_long, aes(x= long, y= lat, 
            size = numero_risposte, colour=differenze_durante_prima), 
            shape=20, stroke=FALSE) +
  scale_size_continuous(name="Numero risposte", range=c(1,12), breaks = mybreaks) +
  #scale_color_viridis(name="Cambio in umore\nDurante-Prima", breaks = c(-3.6,0,3.6), labels = c("Diminuito","Invariato", "Cresciuto")) +
  scale_colour_gradient2(low = "#5e3c99", high = "#e66101", mid = "white", breaks = c(-3.6,0,3.6), labels = c("Diminuito","Invariato", "Cresciuto"))+
  theme_void() + coord_map() + 
  geom_text_repel(data=risposte_by_city %>% arrange(numero_risposte) %>% tail(20), 
                  aes(x=long, y=lat, label=toupper(comune)), size=2,
                  segment.size = 0.2)+
  ggtitle("Risposte per provincia") +
  theme(
      text = element_text(color = "#22211d"),
      plot.background = element_rect(fill = "white", color = NA), 
      panel.background = element_rect(fill = "white", color = NA), 
      legend.background = element_rect(fill = "white", color = NA),
      plot.title = element_text(size= 16, hjust=0.1, color = "#4e4d47")) +
  facet_wrap(~classe_umore)
  #theme(legend.position = "bottom")
```


```{r barplots-classe-umore, fig.height=10, fig.width=10}
risposte_umore <- risposte_numeric_long %>%
  group_by(classe_umore, comune) %>%
  summarise(media = mean(differenze_durante_prima)) %>%
  mutate(comune = fct_reorder(comune, media))

ggplot(risposte_umore, aes(x = comune, y = media, fill = media)) + 
  geom_bar(stat= "identity")+
  facet_wrap(~classe_umore) +
  scale_fill_gradient2(low = "#5e3c99", high = "#e66101", mid = "white", 
                       limits = c(-4,4), breaks = c(-3.6,0,3.6), labels = c("Diminuito","Invariato", "Cresciuto"))+
  theme_classic() +
  coord_flip() +
  labs(x = "Comune", y = "Media differenza di umore Durante-Prima per comune")
```


## Umori singoli uno per uno

```{r barplots-umori-singoli, fig.height=10, fig.width=10}
risposte_umore <- risposte_numeric_long %>%
  group_by(Umore, regione) %>%
  summarise(media = mean(differenze_durante_prima)) %>%
  mutate(comune = fct_reorder(regione, media),
         Umore = factor(Umore, levels = c(positive, negative))) 

ggplot(risposte_umore, aes(x = regione, y = media, fill = media)) + 
  geom_bar(stat= "identity")+
  facet_wrap(~Umore) +
  scale_fill_gradient2(low = "#5e3c99", high = "#e66101", mid = "white", 
                       limits = c(-4,4), breaks = c(-3.6,0,3.6), labels = c("Diminuito","Invariato", "Cresciuto"))+
  theme_classic() +
  coord_flip() +
  labs(x = "Regione", y = "Media") + ggtitle("Media della differenza di umore Durante-Prima per regione")
```
